refactor: standardize parser modules with common XML utilities
Create a shared `common::xml_utils` module to standardize XML parsing
utilities across PubMed and PMC parsers, following the parser design
philosophy of using module functions instead of empty structs.

Changes:
- Create src/common/xml_utils.rs with shared XML utility functions
- Move strip_inline_html_tags from pubmed/parser/preprocessing.rs
- Move all XML utilities from pmc/parser/xml_utils.rs to common module
- Update both parsers to re-export from common::xml_utils
- Add comprehensive tests for all XML utility functions

Benefits:
- Eliminates code duplication between parsers
- Provides single source of truth for XML utilities
- Maintains backward compatibility through re-exports
- All existing tests pass (280 library + 35 parser tests)
