+++ pubmed-client/src/common/mod.rs
@@ -1,6 +1,7 @@
-//! Common data structures shared between PubMed and PMC modules
+//! Common data structures and utilities shared between PubMed and PMC modules
 
 pub mod models;
+pub mod xml_utils;
 
 // Re-export common types
 pub use models::{format_author_name, Affiliation, Author};
